Add loan approval and rejection functionality
Achieve target by additions to modules below
- Add target loan availability check and return of appropriate response
if unfound
- Add admin specified status value validity check and appropriate
response if not
- Status update implementation by calling the approve method of the
target loan with the desired approval status value
- Additionally, add a method-filterRepr to get the desired
representation of the response object
- Activate approval route related tests in the loan tests' moule

To test this feature manually, clone and install a local copy of this
application with npm install command or use the heroku version. If using
a local copy, your base url will be http://localhost:3000 while the
heoku base url https://quickcredit-anguandia.herokuapp.com
- Create a loan using the endpoint(url extension) /loans and enter the
loan [body](https://github.com/Anguandia/quickcredit/pull/26) values on
(postman)[]
- Send an approval request using the patch method on end point
/loans/1 (the system sequentially assigns loans from 1) and request body
having status as any thing other than approved and rejected. Notice the
status validation message in the response
- Repeat the approval attemp with an id other than 1 and notice the not
found response
- Repeat the same, using id 1 and status either approved or rejected and
notice the response
- Additionally notice the the success response object is a summarized
copy of a full loan object, testament to the working of the filterRepr
method
- Also notice that successful approval will immediately credit the loan
amount and reflect a balance inclussive of interest not the initial zero
when the loan was created

Shortcommings
- In the current state, both approving and rejecting the loan will get
the loan credited with the application amount, a bug to be fixed next
- The credit loan method currently can only be accessed on the approval
route. Will be extricated to full independence in the complete
application

[finishes #165846397 #165845800]